rename network::message to network::msg
